Added Types Limits and Codes Enums to improve readability
at the start of a match the game was initialized with a int gametype
that came out of nowhere., now the GameCode consists of a whole phrase
e.g. HOLDEM_NO_LIMIT_IS_TOURNAMENT, which indicates what kind of game to
start in TexasHoldem's main() Method.

GameLimit and GameType help setup the game and remove some private final
int class variables in MatchPlayer